requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
google-search-results>=2.4.2
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
